backend/requirements.txt
fastapi
uvicorn
python-multipart
spacy
scikit-learn
pandas
PyPDF2
python-docx
nltk
pydantic
python-dotenv
